Day 22 - part 2
Find the least amount of mana that can be spent and still wins the fight
on hard difficulty.
Also fix bugs in the part 1 implementation that worked correctly on easy
difficulty, but didn't on hard difficulty.
Add ability to print gameplay as per the problem description's examples
to aid debugging.