Solve day 2 for 2025 (#2701)
* Solve day 2 part 1

Solve part 1 of the puzzle for day 2 of 2025.

* Solve day 2 part 2

Solve part 2 of the puzzle for day 2 of 2025.

* Fix typo

Fix typo in comment.

* Make Digits() use bytes

Use the smallest data type for a single digit number.

* Optimize Math.Digits()

Return `byte[]` (or `ROS<byte>`) and avoid use of intermediate list.

* Remove casts

Use signed values instead of unsigned.

* Add generic Digits<T>()

Add a variant of `Digits<T>()` for any `IBinaryInteger<T>`. Previous overloads retained for performance to avoid virtual dispatch.

* Simplify code

Use `Bifurcate()` to avoid manual iteration.

* Refactor FromDigits()

- Re-use input parameter instead of temporary variable.
- Use `T.DivRem()`.