Partially solve day 10 for 2025 (#2726)
* Solve day 10 part 1

Solve part 1 of the puzzle for day 10 of 2025.

Needs to be made more efficient as it's too slow.

* Optimise solution

Avoid some work, but still pretty slow.

* More refactoring and optimization

- Remove redundant collection.
- Observe `CancellationToken`.
- Rename variable.
- Shortcut earlier where possible.

* Refactor parsing

Refactor parsing and parse joltage ready for part 2.

* Speed-up solution

Shortcut at one value less to avoid redundant work.

* Enable puzzle

No longer prohibitively slow.

* Partially solve day 10 part 2

Partially solve part 2 of the puzzle for day 10 of 2025.

Gets the correct result for the example, but is too slow to determine the result for the input.

* Use A*

Use A* to determine the minimum steps to turn on.

* Use A*

Use A* to solve part 2 as well.

* Disable slow puzzles

Disable the full solution tests for part 2 as it's too slow.

* Address review comments

- Use `ImmutableArray<T>`.
- Fix typos.
- Use correct return type.